web/requirements/base.txt
Django==1.11
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==2.0.6
requests>=2.3.0
django-model-utils
six
Pillow
celery
django-celery
django-analytical
django-s3direct
mercadopago
social-auth-app-django==1.0.0
django-cities-light


